Yet another improvement on s stats
This code is faster and the stats are better.

Instead of retraining s an additional time during the main loop (where 
stats are not collected). The additional s training has been moved 
outside and using all the data (so to have a clear picture of how good 
the current network is able to predict s).