Changes interpretation epoch number specified in the schedule parameter
Before this update the number of epochs specified in the schedule parameters
were interpreted as "number of additional epochs starting from the  current
epoch".

Now it is interpreted as: number of epochs to be reached. The difference
is that now if the parameter says "1000" and we already computed 500
epochs in another run, the script will only perform additiona 500
epochs.